Use `https` for Artifacts submodule, change publish step to save existing commit hash (#72)
Soundtrack of this PR: [link to song that really fits the mood of this PR]()

### Motivation

Im changing the submodule remote URL to be `https`, Im hoping this solves for the `git` issues Im seeing in CICD. My theory is the deploy keys only have access to one repo, and even tho the Artifacts submodule is public, asking it to use SSH forces it to use the deploy key which causes failure. 

Im also changing the publish step. Before after merge to master the same jobs would run + the publish job. Problem was the generative jobs were creating new commits but then also failing leaving the repo in a bad-state. The fix is to not run the generative jobs, they're not needed. Instead we copy the latest Artifacts commit to a new branch, this allows us to delete old branches after PR merge while keeping the latest one around via its new branch name. 

### Future Work
* Re-consider `git lfs` has been annoying.